feat: convert 4 included_stub functions (scene wrappers + rotation matrix)
- func_08014440: main_menu scene wrapper with D_03006518 conditional + set_pause_beatscript_scene + func_0800C7A4
- func_080117A8: main_menu sprite setup with sprite_set_anim_cel + sprite_set_x_y via asm volatile BL (s16 callee-signature trap), D_083A9CE0 position table lookup
- func_080149BC: main_menu scene init with func_0800A240 + func_08005600 + OR 0x10 into gCurrentSceneData+0xDE
- func_08001C74: code_08001a70 rotation matrix using __divsi3 via asm volatile BL (argument reorder trap)

Key patterns:
- asm volatile BL for __divsi3 prevents compiler from reordering R0/R1 args (0x10000 immediate vs angle variable)
- asm volatile "mov r4, r2" prevents MOV+LSL collapse into 3-operand LSL
- Fixed extern type conflicts in asm_080117fc.c and asm_080118e0.c (func_080117A8 u8→s32)
- Documented __divsi3 argument reorder trap and MOV+LSL collapse trap in pattern library

Metrics: 1055 decomp files (873 standalone_tu + 182 included_stub), 22.65% matched functions

+++ src/decomp/asm_080117fc.c
@@ -2,7 +2,7 @@
 #include "global.h"
 #include "src/scenes/main_menu.h"
 
-extern void func_080117A8(u8);
+extern void func_080117A8(s32);
 extern void func_0800C77C(u32);
 extern void func_0800C7A4(s32);
